+++ src/test/java/com/github/f4b6a3/uuid/factory/function/impl/WindowsTimeFunctionTest.java
@@ -4,19 +4,19 @@ import org.junit.Test;
 
 import static org.junit.Assert.*;
 
-public class DefaultTimeFunctionTest {
+public class WindowsTimeFunctionTest {
 
 	private static final int DEFAULT_LOOP_MAX = 1_000_000;
 
 	@Test
 	public void testGetTimestampTicks() {
 		// 1ms = 10,000 ticks
-		DefaultTimeFunction supplier = new DefaultTimeFunction();
+		WindowsTimeFunction supplier = new WindowsTimeFunction();
 		for (int i = 0; i < DEFAULT_LOOP_MAX; i++) {
 			long m1 = System.currentTimeMillis() * 10000L;
 			long ts = supplier.getAsLong();
-			// TS can be 20,000 ticks ahead due to counter shift
-			long m2 = System.currentTimeMillis() * 10000L + 20000L;
+			// TS can be 600,000 ticks ahead due to time granularity and counter shift
+			long m2 = System.currentTimeMillis() * 10000L + 600000L;
 			assertTrue("The current timstamp millisecond is incorrect", ts >= m1 && ts <= m2);
 		}
 	}
@@ -24,12 +24,12 @@ public class DefaultTimeFunctionTest {
 	@Test
 	public void testGetTimestampMillisecond() {
 		// 1ms = 10,000 ticks
-		DefaultTimeFunction supplier = new DefaultTimeFunction();
+		WindowsTimeFunction supplier = new WindowsTimeFunction();
 		for (int i = 0; i < DEFAULT_LOOP_MAX; i++) {
 			long m1 = System.currentTimeMillis();
 			long ts = supplier.getAsLong() / 10000L;
-			// TS can be 1ms ahead due to counter shift
-			long m2 = System.currentTimeMillis() + 1;
+			// TS can be 60ms ahead due to time granularity and counter shift
+			long m2 = System.currentTimeMillis() + 60;
 			assertTrue("The current timstamp millisecond is incorrect", ts >= m1 && ts <= m2);
 		}
 	}
@@ -37,7 +37,7 @@ public class DefaultTimeFunctionTest {
 	@Test
 	public void testGetTimestampMonotonicity() {
 		long lastTs = 0;
-		DefaultTimeFunction supplier = new DefaultTimeFunction();
+		WindowsTimeFunction supplier = new WindowsTimeFunction();
 		for (int i = 0; i < DEFAULT_LOOP_MAX; i++) {
 			long ts = supplier.getAsLong();
 			assertTrue("The current timstamp should be greater than the last one", ts > lastTs);