+++ src/test/java/com/github/f4b6a3/uuid/factory/rfc4122/TimeBasedFactoryTest.java
@@ -6,7 +6,6 @@ import com.github.f4b6a3.uuid.UuidCreator;
 import com.github.f4b6a3.uuid.factory.UuidFactoryTest;
 import com.github.f4b6a3.uuid.factory.function.ClockSeqFunction;
 import com.github.f4b6a3.uuid.factory.function.NodeIdFunction;
-import com.github.f4b6a3.uuid.factory.rfc4122.TimeBasedFactory;
 import com.github.f4b6a3.uuid.util.UuidTime;
 import com.github.f4b6a3.uuid.util.UuidUtil;